Bugfix/1.1

+++ app/src/main/java/com/concordium/wallet/data/walletconnect/Payload.kt
@@ -3,7 +3,7 @@ package com.concordium.wallet.data.walletconnect
 data class Payload(
     val address: ContractAddress,
     val amount: String,
-    var maxEnergy: Int,
+    var maxEnergy: Long,
     val message: String,
     val receiveName: String
 )
